TimeDate::strptime() went up a level!  and returns another param.
so doing (strptime())[-1] is maybe not such a good idea.
and also, the code path that that triggered seems completely
busted and is untested.  added a comment but not touching
one damn thing in here that I don't have to.
so the real news is that this sprintf format:
            '%04d-%02d-%02dT%02d:%02d:%02d.0000000%+03d:%02d',
was generating dates that look like this:
2007-12-31T12:32:00.0000000+00:00
resultin in test failures like this:
t/SOAP/WSDL/XSD/Typelib/Builtin/dateTime.t ............ 1/9
   Failed test 'conversion with timezone'
   at t/SOAP/WSDL/XSD/Typelib/Builtin/dateTime.t line 35.
          got: '2007-08-31T00:32:00.0000000+00:00'
     expected: '2007-08-31T00:32:00-04:00'
looking at the correct strptime() result array field for the
timezone, which is undef (translated to 0) in the test cases
lets us avoid stepping in to that codepath.

resolves FTBFS: test failures with libtimedate-perl 2.3100