allow empty endpoints
Thanks, Andrew Bauer

Build test failed on Fedora 26 Rawhide, but succeeded in previous releases.

The HTTP.t test returned:

Dubious, test returned 255 (wstat 65280, 0xff00)
Failed 4/8 subtests

Accompanied with:

Can't call method "scheme" on an undefined value at /usr/share/perl5/LWP/UserAgent.pm line 135.

Root cause of the issue was the endpoint was not defined. The assumption is an undefined endpoint should not be fatal, and the patch remedies that. If $endpoint is not defined, then set it to be the empty string. It was not investigated why older Fedora distros do not fail on this, but the patch has been verified to be backwards compatible.

resolves #120112: [PATCH] Fix HTTP.pm to allow an undefined endpoint